Initial commit for new API
This API simplifies the programming significantly and hopefully will
speed things up. By implementing amplitudes as a trait rather than a
struct, we should be able to avoid most of the complications that arise
when we pass them around. The downside is that it's now impossible to
implement operator overloads, but I think we'll find that postfix is
good enough for our use cases. Additionally, it's now on the user to
implement the vectorization of functions. This gets rid of a lot of the
pain in the way amplitudes are resolved, since we can make functions
that don't return anything and do all of the work of adding themselves
to datasets.