Complete rewrite of basically everything.
Sorry, it just works better this way. The philosophy here is that it's
way easier to have the user implement their own loop over the events
because we don't have to convert columnar data into some structure which
is unknown at runtime. We could implement a standard format for input
files, but that limits us to hard-code that standard, and makes it
difficult to create backward-compatible changes if we wanted to add
additional fields. The K-Matrix amplitude is still untested, but just
exists now to demonstrate implementing a node that is "self-dependent".